refactor: extract CreateDashboardButton and NavigationMenuDropdownButton components
