+++ src/components/DashboardsBar/NavigationMenu/index.js
@@ -1,2 +1,1 @@
-export { IconNavigation } from './IconNavigation.js'
-export { NavigationMenu } from './NavigationMenu.js'
+export { NavigationMenuDropdownButton } from './NavigationMenuDropdownButton.js'
