+++ src/components/DashboardsBar/NavigationMenu/styles/NavigationMenuDropdownButton.module.css
@@ -1,0 +1,5 @@
+@media only screen and (max-width: 480px) {
+    .navMenuButtonText {
+        display: none;
+    }
+}
