+++ src/components/DashboardsBar/styles/CreateDashboardButton.module.css
@@ -1,0 +1,5 @@
+@media only screen and (max-width: 480px) {
+    .createDashboardButton {
+        display: none;
+    }
+}
